refactor: update view visualization container structure
