+++ client/src/containers/ViewVisualizationContainer/helpers/getVisualizationHelper.js
@@ -1,0 +1,4 @@
+const getVisualization = (visualizations, id) =>
+  visualizations.filter((visualization) => visualization.id === Number(id))[0];
+
+export default getVisualization;
