+++ client/src/containers/ViewVisualizationContainer/helpers/index.js
@@ -1,0 +1,13 @@
+import getVisualization from './getVisualizationHelper';
+import { getVisualizationComponent, getVisualizationSettings, getVisualizationIcon } from './visualizationDataHelper';
+import { checkIsVisualizationNew, createDataSample, createNewVisualization } from './createVisualizationHelper';
+
+export {
+  getVisualization,
+  getVisualizationComponent,
+  getVisualizationSettings,
+  getVisualizationIcon,
+  checkIsVisualizationNew,
+  createDataSample,
+  createNewVisualization,
+};
