+++ client/src/containers/ViewVisualizationContainer/helpers/initConfigsHelper/index.js
@@ -1,0 +1,5 @@
+import createInitBarChartConfig from './barChartConfig';
+import createInitLineChartConfig from './lineChartConfig';
+import createInitTableConfig from './tableConfig';
+
+export { createInitBarChartConfig, createInitLineChartConfig, createInitTableConfig };
